# Conflicts:
#	user-service/src/main/java/com/harbor/secondHand/user/service/impl/IbalanceServiceImpl.java
#	user-service/src/main/java/com/harbor/secondHand/user/service/impl/UserServiceImpl.java

+++ user-service/src/main/java/com/harbor/secondHand/user/domain/dto/UserMessageDTO.java
@@ -12,7 +12,7 @@ public class UserMessageDTO {
     private String avatar;
     private String phone;
     private String email;
-    private BigDecimal balance;
+    private Integer balance;
     private Integer creditScore;
     private LocalDateTime updateTime;
     private String operationType;
